test(research): split PDF line joining family
